Fixed bug in nlp.R to restrict by tokenNames vector explicitly
The code provided was not completely restricting by the tokenNames vector names for some reason. It would allow more fields to be added, and when a sentence resulted in more fields in the df dataframe to be appended to the out$tokens dataframe, the code would throw an error when implementing rbind. This was especially prevalent with certain numbers. The change committed eliminates that.